написал все функции
